% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-to-mat2.R
\name{basis.integral}
\alias{basis.integral}
\title{Calcul de l'intégrale des fonctions de base}
\usage{
basis.integral(t0, tf, basis.beta, basis.X)
}
\arguments{
\item{t0}{the initial time observation}

\item{tf}{the final time observation}

\item{basis.beta}{the order of the splines basis}

\item{basis.X}{the number of functional predictors}
}
\value{
Matrix that compute the integral of basis functions
}
\description{
Calcul de l'intégrale des fonctions de base
}
