% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gated_features.R
\name{gated.features}
\alias{gated.features}
\title{Function that build the design matrix of the gated network model}
\usage{
gated.features(t.mat, nbasis, n.order, X_fd.list)
}
\arguments{
\item{t.mat}{the matrix of time with the corresponded values of Y.mat}

\item{nbasis}{the number of basis for functional parameters}

\item{n.order}{the order of the splines basis}

\item{X_fd.list}{the list of functional predictors}
}
\value{
Dataframe of the design matrix of the gated network model
}
\description{
Function that build the design matrix of the gated network model
}
