% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-to-mat2.R
\name{get.data2}
\alias{get.data2}
\title{Function that build the design matrix of the linear integral model}
\usage{
get.data2(X_fd.list, Y.mat, t.mat, nbasis, n.order = 4)
}
\arguments{
\item{X_fd.list}{the list of functional predictors}

\item{Y.mat}{the matrix response with of n rows and m = max m_i columns with na values after the end of the observation for each sample}

\item{t.mat}{the matrix of time with the corresponded values of Y.mat}

\item{nbasis}{the number of basis for functional parameters}

\item{n.order}{the order of the splines basis}
}
\value{
Matrix that can be use to build the design matrix of the integral model
}
\description{
Function that build the design matrix of the linear integral model
}
