% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-to-mat1.R
\name{my_funmat1}
\alias{my_funmat1}
\title{Categorial variable for taking the repeated observation of the response}
\usage{
my_funmat1(t.mat, Y.mat)
}
\arguments{
\item{t.mat}{the matrix of time with the corresponded values of Y.mat}

\item{Y.mat}{the matrix response with of n rows and m = max m_i columns with na values after the end of the observation for each sample}
}
\description{
Categorial variable for taking the repeated observation of the response
}
