% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-to-mat1.R
\name{my_funmat2}
\alias{my_funmat2}
\title{Functional representation as matrix : Functionnal covariates}
\usage{
my_funmat2(X_fd.list, basis.beta, t.mat, Y.mat)
}
\arguments{
\item{X_fd.list}{the list of functional predictors}

\item{basis.beta}{the (B-splines) for functional parameters}

\item{t.mat}{the matrix of time with the corresponded values of Y.mat}

\item{Y.mat}{the matrix response with of n rows and m = max m_i columns with na values after the end of the observation for each sample}
}
\value{
Matrix that can be use to build the design matrix of the integral model
}
\description{
Functional representation as matrix : Functionnal covariates
}
