% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func-to-mat2.R
\name{my_penmat2}
\alias{my_penmat2}
\title{Penalty matrix for the integral model}
\usage{
my_penmat2(obs.grid, L.beta, n.order, d)
}
\arguments{
\item{obs.grid}{the time observation grid}

\item{L.beta}{the number of basis functions of parameters (an integer)}

\item{n.order}{the order of the splines basis}

\item{d}{the number of functional predictors}
}
\value{
Matrix that can be use to penalize the integral model
}
\description{
Penalty matrix for the integral model
}
