% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penffr2.R
\name{penffr2}
\alias{penffr2}
\title{Functional integral model}
\usage{
penffr2(
  Y.mat,
  t.mat = "all.ok",
  X_fd.list,
  X.scal = data.frame(),
  nbasis = "all.ok",
  n.order = 4,
  pen = T,
  weights = NULL
)
}
\arguments{
\item{Y.mat}{the matrix response with of n rows and m = max m_i columns with na values after the end of the observation for each sample}

\item{t.mat}{the matrix of time with the corresponded values of Y.mat}

\item{X_fd.list}{the list of functional predictors}

\item{X.scal}{the data frame of the non functional predictors}

\item{nbasis}{the number of basis for functional parameter (an integer)}

\item{n.order}{the order of the splines basis}

\item{pen}{the boolean value that indicates if we want to use the penalization}
}
\value{
a list that containt
\itemize{
\item \code{model}: the fitting model object
\item \code{beta.fd}: the estimated functional parameters.
\item \code{beta.scal}: the estimated scalar parameters
\item \code{lambda}: the optimal penalized parameter
\item \code{n.grid}: the length of time observation (useful for the predict function)
}
}
\description{
Functional integral model
}
