% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penffr2_mix.R
\name{penffr2_mix}
\alias{penffr2_mix}
\title{Functional Mixture-of-Experts model for integral model}
\usage{
penffr2_mix(
  Y.mat,
  t.mat = "all.ok",
  X_fd.list,
  X.scal = data.frame(),
  nbasis = "all.ok",
  K.comp = "BIC",
  pen = T
)
}
\arguments{
\item{Y.mat}{the matrix response with of n rows and m = max m_i columns with na values after the end of the observation for each sample}

\item{t.mat}{the matrix of time with the corresponded values of Y.mat}

\item{X_fd.list}{the list of functional predictors}

\item{X.scal}{the data frame of the non functional predictors}

\item{nbasis}{the number of basis for functional parameter (an integer)}

\item{K.comp}{the number of fixed component to have or "BIC" to select the best model using BIC criterion}

\item{pen}{the boolean value that indicates if we want to use the penalization}

\item{n.order}{the order of the splines basis}

\item{K.min}{minimum number of components to consider}

\item{K.max}{maximum number of components to consider}

\item{K.mixture}{the number of component mixture}
}
\value{
a list that containt
\itemize{
\item \code{model}: the fitting model object
\item \code{beta.fd}: the estimated functional parameters.
\item \code{beta.scal}: the estimated scalar parameters
\item \code{lambda}: the optimal penalized parameter
\item \code{n.grid}: the length of time observation (useful for the predict function)
}
}
\description{
Functional Mixture-of-Experts model for integral model
}
