% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PenFFR.R
\name{pred.PenFFR}
\alias{pred.PenFFR}
\title{Prediction of Functional model}
\usage{
pred.PenFFR(
  model,
  newX_fd.list = list(),
  newX.scal = data.frame(),
  t.mat = "all.ok",
  K.mixture = "No",
  model.type = "concurrent"
)
}
\arguments{
\item{model}{the functional model}

\item{newX_fd.list}{the list of functional predictors}

\item{newX.scal}{the data frame of the non functional predictors}

\item{t.mat}{the matrix of time on where we want the prediction}
}
\value{
the prediction of functional model
}
\description{
Prediction of Functional model
}
