% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penffr1.R
\name{pred.penffr1}
\alias{pred.penffr1}
\title{Prediction of Functional concurrent model}
\usage{
pred.penffr1(
  model,
  newX_fd.list = list(),
  newX.scal = data.frame(),
  t.mat = "all.ok"
)
}
\arguments{
\item{model}{the functional model}

\item{newX_fd.list}{the list of functional predictors}

\item{newX.scal}{the data frame of the non functional predictors}

\item{t.mat}{the matrix of time on where we want the prediction}
}
\value{
the prediction of functional concurrent model
}
\description{
Prediction of Functional concurrent model
}
