% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penffr1_mix.R
\name{pred.penffr1_mix}
\alias{pred.penffr1_mix}
\title{Prediction of Functional Mixture-of-Experts model for concurrent model}
\usage{
pred.penffr1_mix(
  model,
  newX_fd.list = list(),
  newX.scal = data.frame(),
  t.mat = "all.ok"
)
}
\arguments{
\item{model}{the functional Mixture-of-Experts model}

\item{newX_fd.list}{the list of functional predictors}

\item{newX.scal}{the data frame of the non functional predictors}

\item{t.mat}{the matrix of time on where we want the prediction}
}
\value{
the prediction of functional Mixture-of-Experts model for concurrent model
}
\description{
Prediction of Functional Mixture-of-Experts model for concurrent model
}
