% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penffr2_mix.R
\name{pred.penffr2_mix}
\alias{pred.penffr2_mix}
\title{Prediction of Functional Mixture-of-Experts model for integral model}
\usage{
pred.penffr2_mix(
  model,
  newX_fd.list = list(),
  newX.scal = data.frame(),
  t.mat = "all.ok"
)
}
\arguments{
\item{model}{the functional Mixture-of-Experts model}

\item{newX_fd.list}{the list of functional predictors}

\item{newX.scal}{the data frame of the non functional predictors}

\item{t.mat}{the matrix of time on where we want the prediction}
}
\value{
the vector of predictions
}
\description{
Prediction of Functional Mixture-of-Experts model for integral model
}
